requirements.txt
streamlit>=1.33.0
speechrecognition
pydub
python-docx
